requirements.txt
flask
flask-cors
google-play-scraper
tensorflow-cpu
pandas
numpy
joblib
gunicorn